Split methods on Node subtypes between regular Nodes and Nodes with diagnostic abilities (like the Imp)
